Balance in new nodes to avoid "thundering herd".
Adds a new configuration variable "ramp_up_time" that specifies the
number of seconds over which to perform the ramping. It defaults to 60.

A node ramps up linearly over time.  In the future it might be good to
explore other strategies (logarithmic perhaps?) or make it configurable.

Ramp up period is configurable for now but in the future should possibly
take into account the load on the cluster and the known limits
(connections, etc.) each node can take.