Persist node start_time when config is reloaded.
When the configuration is reloaded, we persist the start_time of any
existing nodes in a cluster so that they are not re-ramped up into the
cluster since they are presumed to already be ramped up or in the
process of ramping up.

Refactored Node to take a hash initializer.