Add downpage node if cluster health is zero.
When all of the nodes in a cluster report zero-health, add a "downpage"
node on the loopback interface to the IPVS cluster that serves static
downfile content or a maintentance page. Service of this downfile is
provided by a service other than Big Brother. One the cluster health
recovers, the downpage node is removed.