First Pass - Merge acitve/active and active/passive clusters into cluster
A lot of cleanup is still needed, because the main cluster is still a
mess, and we can probably safely remove the ActivePassiveCluster and
ActiveActiveCluster classes, but this _should_ be ready to be used.

The AA and AP specs should also probably be merged, but at present
they're still separate because it was the simplest way to move forward.
The only difference is that they now are testing against the Cluster
class (because ClusterFactory now only returns Clusters).