INF-985: Start remote nodes in synchronize!
This fixes a bug where an AA cluster will never start remote LB nodes if
there are no local nodes defined. This is a fairly niche case that we
see when standing up the ldap_primary cluster, which only lives in one
DC.

This change will start the remote nodes explicitly rather than let them
be started as part of a later `monitor_nodes` operation.