Fix init script for bigbro and bump debian release to 0.9.2-2
When the interpreter is set to "/usr/bin/env ruby", as it was in older
versions of the package, the comm name of the process is "ruby".  This
meant that the init script needed to account for that so the process
could be stopped properly.  Now that the interpreter is "/usr/bin/ruby"
the comm name of the process is (arguably more appropriately) set to
"bigbro".  This commit fixes the init script to work with a comm name of
bigbro.