Check cluster mode instaed of backend mode
This was causing errors in the cluster state check when the backend mode
wasn't explicitly set in the config file. Taking a look at the usage of
backend_mode and cluster_mode, it's possible that they may be able to be
condensed into a single value. Will investigate that further.