Match nodes to initial IPVS state when starting/restarting a cluster
We encountered an issue where clusters that were restarted or migrated,
had backends that were set to 0. When the cluster is stopped and then
started, the weights are now invalidated to be set to 0 with our change to
`invalidate_weight!`, but the IPVS weight of the node is still
initialized with a value of 1 causing a mismatch in BB and IPVS weights.

This change ensures that node weights will be updated on the next
monitoring check even if the measured health is zero or if there is a
downpage. It is also important to note (as seen below) that at this
time, we prefer to set the value to 1 on the initialization of a
cluster, even if the measured or configured weight is supposed to be 0.