Default the weight of nodes to 0
Initialize a node with 0 if not given a weight to ensure that a weight
value is never set to nil. Previously, if the weight was not defined,
the node would be added to ipvs without a weight, causing the ipvsadm
execution to fail. This synchronizes the behavior of adding a node to an
existing cluster with how we handle adding a new cluster entirely.