Set a cluster's monitored state to the original cluster's when reloading
When reloading big brother, existing clusters are replaced with a
cluster instance that is instantiated with a monitored value of false.
This means the cluster won't be monitored again until the cluster is
synchronized in a separate action. This change will carry over the state
from an existing cluster which should prevent issues when a cluster
isn't monitored when it should be.