Add a status_detail route for node health reporting
This commit adds a status_detail route that reports the health and
address of each node in the cluster. This is so tools that check service
health can easily determine the number of live nodes and average
weight across nodes, which isn't possible when we only report combined
weight and whether the cluster is running.

This also changes the Travis CI config to update bundler before running,
since outdated versions of bundler that come with system ruby were
causing builds to fail.